Music drop, fixes, improvments, Krobot 2.1.0

+++ src/main/java/fr/litarvan/shenron/command/CommandClearWhere.java
@@ -7,14 +7,33 @@ import fr.litarvan.krobot.util.Dialog;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import javax.inject.Inject;
+import net.dv8tion.jda.core.JDA;
+import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.entities.MessageHistory;
+import org.jetbrains.annotations.NotNull;
 
 public class CommandClearWhere implements CommandHandler
 {
+    @Inject
+    private JDA jda;
+
     @Override
-    public void handle(CommandContext context, Map<String, SuppliedArgument> args) throws Exception
+    public void handle(@NotNull CommandContext context, @NotNull Map<String, SuppliedArgument> args) throws Exception
     {
+        if (!context.getMember().hasPermission(context.getChannel(), Permission.MESSAGE_MANAGE))
+        {
+            context.getChannel().sendMessage(Dialog.error("Non-autorisé", "Vous n'avez pas la permission de supprimer les messages")).queue();
+            return;
+        }
+
+        if (!context.getGuild().getMember(jda.getSelfUser()).hasPermission(context.getChannel(), Permission.MESSAGE_MANAGE))
+        {
+            context.getChannel().sendMessage(Dialog.error("Non-autorisé", "Shenron n'a pas la permission de supprimer les messages")).queue();
+            return;
+        }
+
         context.getMessage().delete().submit().get();
 
         List<Message> messages = context.getChannel().getHistory().retrievePast(100).complete();
@@ -33,7 +52,7 @@ public class CommandClearWhere implements CommandHandler
 
         if (from == null)
         {
-            context.getChannel().sendMessage(Dialog.info("Erreur", "Impossible de trouver le message")).queue();
+            context.sendMessage(Dialog.info("Erreur", "Impossible de trouver le message"));
             return;
         }
